use models dir instead of file to remove very long models.py
