+++ cms/snippets/models/__init__.py
@@ -1,0 +1,6 @@
+from wagtail.wagtailsnippets.models import register_snippet
+from .categories import Category
+from .navigation_menus import NavigationMenu
+
+register_snippet(Category)
+register_snippet(NavigationMenu)
